fixed createdAt type

+++ OpenAI.SDK/ObjectModels/ResponseModels/ChatCompletionCreateResponse.cs
@@ -38,11 +38,7 @@ public record ChatCompletionCreateResponse : BaseResponse, IOpenAIModels.IId, IO
     [JsonPropertyName("service_tier")]
     public string? ServiceTier { get; set; }
 
-    /// <summary>
-    ///     The Unix timestamp (in seconds) of when the chat completion was created.
-    /// </summary>
-    [JsonPropertyName("created")]
-    public int CreatedAt { get; set; }
+    [JsonPropertyName("created")] public long CreatedAt { get; set; }
 
     /// <summary>
     ///     A unique identifier for the chat completion.
