+++ OpenAI.SDK/ObjectModels/SharedModels/FileResponse.cs
@@ -12,16 +12,8 @@ public record FileResponse : BaseResponse, IOpenAIModels.IId, IOpenAIModels.ICre
     public string FileName { get; set; }
 
     public UploadFilePurposes.UploadFilePurpose PurposeEnum => UploadFilePurposes.ToEnum(Purpose);
-
-    [JsonPropertyName("purpose")]
-    public string Purpose { get; set; }
-
-    [JsonPropertyName("status")]
-    public string Status { get; set; }
-
-    [JsonPropertyName("created_at")]
-    public int CreatedAt { get; set; }
-
-    [JsonPropertyName("id")]
-    public string Id { get; set; }
+    [JsonPropertyName("purpose")] public string Purpose { get; set; }
+    [JsonPropertyName("status")] public string Status { get; set; }
+    [JsonPropertyName("created_at")] public long CreatedAt { get; set; }
+    [JsonPropertyName("id")] public string Id { get; set; }
 }
